location consuming store information
