+++ src/App.js
@@ -9,18 +9,7 @@ import WeatherContainer from './components/Weather/WeatherContainer';
 
 export default class components extends Component {
 
-  dataBuilder = (d) => {
-    let months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
-    let days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"]
-    let day = days[d.getDay()];
-    let date = d.getDate();
-    let month = months[d.getMonth()];
-    let year = d.getFullYear();
-    return `${day} ${date} ${month} ${year}`
-  }
-
   render() {
-
     return (
       <Provider store={store}>
         {/* <div className={typeof this.state.temperature != "undefined" ? ((this.state.temperature > 18) ? 'app warm' : 'app') : 'app'}> */}
@@ -28,7 +17,7 @@ export default class components extends Component {
           <main>
             <SearchContainer />
             <div>
-              <LocationContainer date={this.dataBuilder(new Date())} />
+              <LocationContainer />
               <WeatherContainer />
             </div>
           </main>
