+++ src/components/Location/LocationContainer.js
@@ -1,18 +1,36 @@
 import React from 'react';
 import Location from './Location';
+import { connect } from 'react-redux';
 
 class LocationContainer extends React.Component {
+    dataBuilder = (d) => {
+        let months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
+        let days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"]
+        let day = days[d.getDay()];
+        let date = d.getDate();
+        let month = months[d.getMonth()];
+        let year = d.getFullYear();
+        return `${day} ${date} ${month} ${year}`
+    }
 
     render() {
+        const { country, city } = this.props;
         return (
             <Location
-                // city={this.props.localState.city}
-                // location={this.props.localState.country}
-                // date={this.props.date}
+                city={city}
+                location={country}
+                date={this.dataBuilder(new Date())}
             />
         )
     }
 };
 
-export default LocationContainer;
+const mapStateToProps = state => (
+    {
+        country: state.cityWeather[0] && state.cityWeather[0].sys.country,
+        city: state.cityWeather[0] && state.cityWeather[0].name
+    });
+
+export default connect(mapStateToProps)(LocationContainer);
+
